+++ openquake/man/general_utils/make_gmm_list.py
