+++ openquake/man/tests/general_utils/data_csv_output_test/case_8/ssm/source_model.xml
