+++ openquake/man/tests/general_utils/data_csv_output_test/ruptures.csv
